Fix GROUP BY and SET parsers
A GROUP BY expression with integer literals refer to the indexed columns.
After changing the literals from int64 to int32, the GROUP BY parser stopped
recognizing the numbers as integers and thought they were actual names of
columns.

A SET expression detects the ON and OFF keywords and resolves them to 1 and
0, respectively. These integers also needed to be converted to int32.

Signed-off-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>